Move state check func and classes to seperate file
